moved tesseract to converters
