+++ featurex/tests/test_extractors.py
@@ -8,7 +8,6 @@ from featurex.extractors.audio import STFTExtractor, MeanAmplitudeExtractor
 from featurex.extractors.image import (BrightnessExtractor,
                                         SharpnessExtractor,
                                         VibranceExtractor,
-                                        TesseractExtractor,
                                         SaliencyExtractor)
 from featurex.extractors.video import DenseOpticalFlowExtractor
 from featurex.extractors.api import (IndicoAPIExtractor,
@@ -106,14 +105,6 @@ def test_vibrance_extractor():
     val = stim.extract([VibranceExtractor()])
     color = val.data['VibranceExtractor'].data['avg_color']
     assert np.isclose(color,1370.65482988)
-
-def test_tesseract_extractor():
-    pytest.importorskip('pytesseract')
-    image_dir = join(get_test_data_path(), 'image')
-    stim = ImageStim(join(image_dir, 'button.jpg'))
-    ext = TesseractExtractor()
-    output = ext.transform(stim).data['text']
-    assert output == 'Exit'
 
 def test_saliency_extractor():
     pytest.importorskip('cv2')
